requirements.txt
torch
transformers
opencv-python
ultralytics
pillow
